+++ Services/SecMaster/V2/Model/ListDataclassRequest.php
@@ -34,8 +34,8 @@ class ListDataclassRequest implements ModelInterface, ArrayAccess
     protected static $openAPITypes = [
             'contentType' => 'string',
             'workspaceId' => 'string',
-            'offset' => 'float',
-            'limit' => 'float',
+            'offset' => 'int',
+            'limit' => 'int',
             'name' => 'string',
             'businessCode' => 'string',
             'description' => 'string',
@@ -349,7 +349,7 @@ class ListDataclassRequest implements ModelInterface, ArrayAccess
     * Gets offset
     *  偏移量
     *
-    * @return float|null
+    * @return int|null
     */
     public function getOffset()
     {
@@ -359,7 +359,7 @@ class ListDataclassRequest implements ModelInterface, ArrayAccess
     /**
     * Sets offset
     *
-    * @param float|null $offset 偏移量
+    * @param int|null $offset 偏移量
     *
     * @return $this
     */
@@ -373,7 +373,7 @@ class ListDataclassRequest implements ModelInterface, ArrayAccess
     * Gets limit
     *  数据量
     *
-    * @return float|null
+    * @return int|null
     */
     public function getLimit()
     {
@@ -383,7 +383,7 @@ class ListDataclassRequest implements ModelInterface, ArrayAccess
     /**
     * Sets limit
     *
-    * @param float|null $limit 数据量
+    * @param int|null $limit 数据量
     *
     * @return $this
     */
