+++ Services/SecMaster/V2/Model/ListWorkflowsRequest.php
@@ -39,8 +39,8 @@ class ListWorkflowsRequest implements ModelInterface, ArrayAccess
     protected static $openAPITypes = [
             'contentType' => 'string',
             'workspaceId' => 'string',
-            'offset' => 'float',
-            'limit' => 'float',
+            'offset' => 'int',
+            'limit' => 'int',
             'order' => 'string',
             'sortby' => 'string',
             'enabled' => 'bool',
@@ -474,7 +474,7 @@ class ListWorkflowsRequest implements ModelInterface, ArrayAccess
     * Gets offset
     *  偏移量
     *
-    * @return float|null
+    * @return int|null
     */
     public function getOffset()
     {
@@ -484,7 +484,7 @@ class ListWorkflowsRequest implements ModelInterface, ArrayAccess
     /**
     * Sets offset
     *
-    * @param float|null $offset 偏移量
+    * @param int|null $offset 偏移量
     *
     * @return $this
     */
@@ -498,7 +498,7 @@ class ListWorkflowsRequest implements ModelInterface, ArrayAccess
     * Gets limit
     *  数据量
     *
-    * @return float|null
+    * @return int|null
     */
     public function getLimit()
     {
@@ -508,7 +508,7 @@ class ListWorkflowsRequest implements ModelInterface, ArrayAccess
     /**
     * Sets limit
     *
-    * @param float|null $limit 数据量
+    * @param int|null $limit 数据量
     *
     * @return $this
     */
